Rename package
i think i actually spelled it right this time...
