+++ app/build.gradle
@@ -9,7 +9,7 @@ apply plugin: 'kotlin-kapt'
 android {
     compileSdkVersion 28
     defaultConfig {
-        applicationId "com.lambdaschool.empoweredconverstation"
+        applicationId "com.lambdaschool.empoweredconversation"
         minSdkVersion 21
         targetSdkVersion 28
         versionCode 1
