+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-          package="com.lambdaschool.empoweredconverstation">
+          package="com.lambdaschool.empoweredconversation">
 
     <application
             android:allowBackup="true"
@@ -9,14 +9,14 @@
             android:roundIcon="@mipmap/ic_launcher_round"
             android:supportsRtl="true"
             android:theme="@style/AppTheme">
-        <activity android:name="com.lambdaschool.empoweredconverstation.LoginForm">
+        <activity android:name=".LoginForm">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
 
                 <category android:name="android.intent.category.LAUNCHER"/>
             </intent-filter>
         </activity>
-        <activity android:name="com.lambdaschool.empoweredconverstation.MainActivity">
+        <activity android:name=".MainActivity">
         </activity>
     </application>
