+++ app/src/main/java/com/lambdaschool/empoweredconversation/User.kt
@@ -1,0 +1,3 @@
+package com.lambdaschool.empoweredconversation
+
+data class User (val name: String)
