--- app/src/main/java/com/lambdaschool/empoweredconverstation/User.kt
@@ -1,3 +1,0 @@
-package com.lambdaschool.empoweredconverstation
-
-data class User (val name: String)
