+++ app/src/main/res/layout/activity_login_form.xml
@@ -4,7 +4,7 @@
         xmlns:app="http://schemas.android.com/apk/res-auto" xmlns:tools="http://schemas.android.com/tools"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context="com.lambdaschool.empoweredconverstation.LoginForm">
+        tools:context="com.lambdaschool.empoweredconversation.LoginForm">
 
     <androidx.constraintlayout.widget.ConstraintLayout
             android:layout_width="match_parent"
